// src/graphql/generated/getAllGamesAnyPlayers.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: getAllGamesAnyPlayers
// ====================================================

export interface getAllGamesAnyPlayers_getAllGamesAnyPlayers {
  __typename: "Player";
  ID: string | null;
  playerID: string | null;
  name: string | null;
  score: number | null;
  game: string | null;
}

export interface getAllGamesAnyPlayers {
  getAllGamesAnyPlayers: getAllGamesAnyPlayers_getAllGamesAnyPlayers[] | null;
}